fix error when no backup file and error in README.md

+++ DynamicPatcher/Patcher.cs
@@ -104,7 +104,7 @@ namespace DynamicPatcher
             CompilationManager = new CompilationManager(workDir);
 #endif
 
-            if (!copyLogFiles)
+            if (!copyLogFiles && !string.IsNullOrEmpty(backupFileName))
             {
                 File.Delete(backupFileName);
             }
